+++ src/structlog/testing.py
@@ -18,7 +18,7 @@ from typing import Any, Generator, NamedTuple, NoReturn
 
 from ._config import configure, get_config
 from .exceptions import DropEvent
-from .types import EventDict, WrappedLogger
+from .typing import EventDict, WrappedLogger
 
 
 __all__ = ["LogCapture", "capture_logs"]
@@ -30,7 +30,7 @@ class LogCapture:
     Generally you should use `structlog.testing.capture_logs`,
     but you can use this class if you want to capture logs with other patterns.
 
-    :ivar List[structlog.types.EventDict] entries: The captured log entries.
+    :ivar List[structlog.typing.EventDict] entries: The captured log entries.
 
     .. versionadded:: 20.1.0
     """